requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors==0.1
django-cors-headers==3.2.1
django-localflavor==3.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dwollav2==2.0.0
future==0.18.2
idna==2.8
PyJWT==1.7.1
python-stdnum==1.13
pytz==2019.3
requests==2.21.0
shippo==2.0.0
simplejson==3.16.0
sqlparse==0.3.1
urllib3==1.24.3
stripe
django-otp-yubikey
psycopg2
django-filter
plaid-python
django-widget-tweaks
python-decouple
django-crispy-forms==1.9.2
django-formtools==2.2
django-otp==0.9.4
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-two-factor-auth==1.12.1
phonenumbers==8.12.8
qrcode==6.1
six==1.15.0
sqlparse==0.3.1
